data id added to emitter
data is sent alongside a data id (did). When ack is not received, the did is not changed. This way, if the initial transmition fails the gateway will not notice, but if ack fails, the gateway will notice and the data wll not be processed twice. did is a random number from 0 to 254 and the same id cannot be the same for two consecutive different data.