sleepFor and SleepUntil implemented
sleepFor: translates seconds, minutes, hours and days into us, and goes to sleep. Only seconds is a mandatory input.

sleepUntil: gets the epoch time of the given date, the actual one, and substracts both to take the difference in seconds. If given date is in the past, returns error. This function can make the rtc go back up to one second every time that it is called. Therefore, it may be good to refresh the emitter time with new gateway info every now and then