emitter fits to schedule
emitter goes to sleep when an ack is received, or when it detects that the collection time has ended.

Emitter wakes up when the next collection time frame opens.

Main code is too spaghetti. A new function should be added, to know if actual time is before or after a certine schedule time, and another function to manage sleep in all situations.

Maybe the code that determines if collection happens during midnight can be reused with some parameters.

USB1 changed to USB0 since gateway uses ACM0