new functions to control sleep of garbage collection
isInstantBeforeReference(now, reference): says if now is before reference.

sleepManager: calculates the date that the MCU needs to sleep until and calls sleepUntil for that.

outOfScheduleManager: looks if the emitter is active during out of schedule, and calls sleepMaanger if so.

finishedTodaysJobManager: called when gateway has received today's measurement. It simply calls sleepManager, with a few tweaks if midnight has passed or not.

In main:
-spaghetti code cleared through functions above.
-sent_OK variable deleted: emitter reads new value every time, so as to inform gateway with the latest reading.
-traps implemented if something is wrong with the sleep: sleepFor 0xFFFF days (as good as forever).