each bin has its own start and end schedule now.
New file collectionCfg.cpp with a bins table. Each bin has a start and end time (hh,mm).

New function findBin() to get the row of the bins table that contains info of the specific bin_id.

new function tableTimeToSimpleTime() to read the time of bins table and tranfer it to a simple_time struct.

stringTimeTo SimpleTime() is now deprecated.